@@ -1,7 +1,7 @@
 package stream.customratingdialogue;
 
 import android.graphics.drawable.Drawable;
-import android.support.annotation.DrawableRes;
+import androidx.annotation.DrawableRes;
 
 interface SimpleRatingBar {
 
